+++ test/integration/fixtures.ts
@@ -60,7 +60,7 @@ import {
   TestIRToken,
   TestIStRSR,
   RecollateralizationLibP1,
-  FacadeInvariantMonitor,
+  FacadeMonitor,
 } from '../../typechain'
 import {
   Collateral,
@@ -611,7 +611,7 @@ interface DefaultFixture extends RSRAndCompAaveAndCollateralAndModuleFixture {
   facade: FacadeRead
   facadeAct: FacadeAct
   facadeTest: FacadeTest
-  facadeInvariantMonitor: FacadeInvariantMonitor
+  facadeMonitor: FacadeMonitor
   broker: TestIBroker
   rsrTrader: TestIRevenueTrader
   rTokenTrader: TestIRevenueTrader
@@ -689,13 +689,11 @@ const makeDefaultFixture = async (setBasket: boolean): Promise<DefaultFixture> =
   const FacadeTestFactory: ContractFactory = await ethers.getContractFactory('FacadeTest')
   const facadeTest = <FacadeTest>await FacadeTestFactory.deploy()
 
-  // Deploy FacadeInvariantMonitor
-  const FacadeInvariantMonitorFactory: ContractFactory = await ethers.getContractFactory(
-    'FacadeInvariantMonitor'
-  )
+  // Deploy FacadeMonitor
+  const FacadeMonitorFactory: ContractFactory = await ethers.getContractFactory('FacadeMonitor')
 
-  const facadeInvariantMonitor = <FacadeInvariantMonitor>await upgrades.deployProxy(
-    FacadeInvariantMonitorFactory,
+  const facadeMonitor = <FacadeMonitor>await upgrades.deployProxy(
+    FacadeMonitorFactory,
     [owner.address],
     {
       kind: 'uups',
@@ -959,7 +957,7 @@ const makeDefaultFixture = async (setBasket: boolean): Promise<DefaultFixture> =
     facade,
     facadeAct,
     facadeTest,
-    facadeInvariantMonitor,
+    facadeMonitor,
     rsrTrader,
     rTokenTrader,
   }
